Reorganises billing route and controller files
